+++ software/deimos/src/calc/sequence_machine.rs
@@ -193,10 +193,20 @@ impl SequenceLookup {
             return Err("Sequence time entries must be monotonically increasing".to_owned());
         }
 
-        match self.eval_checked(0.0) {
+        // Run the interpolator at the first and last times
+        let first_time = self.time_s.first().ok_or("Empty sequence".to_string())?;
+        match self.eval_checked(*first_time) {
             Ok(_) => Ok(()),
             Err(x) => Err(x),
-        }
+        }?;
+
+        let last_time = self.time_s.last().ok_or("Empty sequence".to_string())?;
+        match self.eval_checked(*last_time) {
+            Ok(_) => Ok(()),
+            Err(x) => Err(x),
+        }?;
+
+        Ok(())
     }
 
     /// Sample the lookup, propagating any errors encountered while assembling or evaluating the interpolator.
@@ -457,7 +467,6 @@ pub struct MachineCfg {
 ///
 /// Unlike most calcs, the names of the inputs and outputs of this calc
 /// are not known at compile-time, and are assembled from inputs instead.
-#[derive(Default)]
 #[cfg_attr(feature = "ser", derive(Serialize, Deserialize))]
 pub struct SequenceMachine {
     /// State transition criteria and other configuration
@@ -478,6 +487,19 @@ pub struct SequenceMachine {
     /// including sequence time and per-run configuration.
     #[cfg_attr(feature = "ser", serde(skip))]
     execution_state: ExecutionState,
+}
+
+impl Default for SequenceMachine {
+    fn default() -> Self {
+        Self {
+            cfg: MachineCfg {
+                entry: "Placeholder".into(),
+                ..Default::default()
+            },
+            sequences: BTreeMap::from([("Placeholder".into(), Sequence::default())]),
+            execution_state: ExecutionState::default(),
+        }
+    }
 }
 
 impl SequenceMachine {